⚡ IMRPOVE: type the styles
