+++ styles/Map.ts
@@ -1,11 +1,11 @@
-import { StyleSheet } from "react-native";
+import { StyleSheet } from "react-native"
 
 export default StyleSheet.create({
   wideMap: {
     height: 200,
+    marginBottom: 10,
     marginLeft: -20,
     marginRight: -20,
     marginTop: 10,
-    marginBottom: 10,
   },
-});
+})
