+++ styles/Splash.ts
