--- styles/common.js
@@ -1,1 +1,0 @@
-export default { flex: { flex: 1 } };
