+++ styles/common.ts
@@ -1,0 +1,1 @@
+export default { flex: { flex: 1 } }
